+++ lib/sidekiq/logging.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 require 'time'
 require 'logger'
+require 'fcntl'
 
 module Sidekiq
   module Logging
@@ -47,7 +48,7 @@ module Sidekiq
     end
 
     def self.logger=(log)
-      @logger = (log ? log : Logger.new('/dev/null'))
+      @logger = (log ? log : Logger.new(File::NULL))
     end
 
     # This reopens ALL logfiles in the process that have been rotated