+++ lib/sidekiq/processor.rb
@@ -111,6 +111,7 @@ module Sidekiq
         end
       end
       sleep(1)
+      nil
     end
 
     def process(work)