+++ lib/sidekiq/util.rb
@@ -58,19 +58,5 @@ module Sidekiq
       end
       arr.clear
     end
-
-    def want_a_hertz_donut?
-      # what's a hertz donut?
-      # punch!  Hurts, don't it?
-      info = Sidekiq.redis {|c| c.info }
-      if info['connected_clients'].to_i > 1000 && info['hz'].to_i >= 10
-        Sidekiq.logger.warn { "Your Redis `hz` setting is too high at #{info['hz']}.  See mperham/sidekiq#2431.  Set it to 3 in #{info[:config_file]}" }
-        true
-      else
-        Sidekiq.logger.debug { "Redis hz: #{info['hz']}.  Client count: #{info['connected_clients']}" }
-        false
-      end
-    end
-
   end
 end