+++ test/test_client.rb
@@ -33,6 +33,23 @@ class TestClient < Sidekiq::Test
       assert_equal 24, jid.size
     end
 
+    it 'allows middleware to stop bulk jobs' do
+      mware = Class.new do
+        def call(worker_klass,msg,q,r)
+          msg['args'][0] == 1 ? yield : false
+        end
+      end
+      client = Sidekiq::Client.new
+      client.middleware do |chain|
+        chain.add mware
+      end
+      q = Sidekiq::Queue.new
+      q.clear
+      result = client.push_bulk('class' => 'Blah', 'args' => [[1],[2],[3]])
+      assert_equal 1, result.size
+      assert_equal 1, q.size
+    end
+
     it 'allows local middleware modification' do
       $called = false
       mware = Class.new { def call(worker_klass,msg,q,r); $called = true; msg;end }
@@ -165,6 +182,18 @@ class TestClient < Sidekiq::Test
       conn.verify
     end
 
+    it 'allows #via to point to same Redi' do
+      conn = MiniTest::Mock.new
+      conn.expect(:multi, [0, 1])
+      sharded_pool = ConnectionPool.new(size: 1) { conn }
+      Sidekiq::Client.via(sharded_pool) do
+        Sidekiq::Client.via(sharded_pool) do
+          CWorker.perform_async(1,2,3)
+        end
+      end
+      conn.verify
+    end
+
     it 'allows #via to point to different Redi' do
       conn = MiniTest::Mock.new
       conn.expect(:multi, [0, 1])
@@ -190,6 +219,44 @@ class TestClient < Sidekiq::Test
       DWorker.sidekiq_options('pool' => ConnectionPool.new(size: 1) { conn })
       Sidekiq::Client.enqueue_in(10, DWorker, 3)
       conn.verify
+    end
+  end
+
+  describe 'Sidekiq::Worker#set' do
+    class SetWorker
+      include Sidekiq::Worker
+      sidekiq_options :queue => :foo, 'retry' => 12
+    end
+
+    def setup
+      Sidekiq.redis {|c| c.flushdb }
+    end
+
+    it 'allows option overrides' do
+      q = Sidekiq::Queue.new('bar')
+      assert_equal 0, q.size
+      assert SetWorker.set(queue: :bar).perform_async(1)
+      job = q.first
+      assert_equal 'bar', job['queue']
+      assert_equal 12, job['retry']
+    end
+
+    it 'handles symbols and strings' do
+      q = Sidekiq::Queue.new('bar')
+      assert_equal 0, q.size
+      assert SetWorker.set('queue' => 'bar', :retry => 11).perform_async(1)
+      job = q.first
+      assert_equal 'bar', job['queue']
+      assert_equal 11, job['retry']
+
+      q.clear
+      assert SetWorker.perform_async(1)
+      assert_equal 0, q.size
+
+      q = Sidekiq::Queue.new('foo')
+      job = q.first
+      assert_equal 'foo', job['queue']
+      assert_equal 12, job['retry']
     end
   end
 end